+++ CHANGELOG.md
@@ -8,6 +8,7 @@
      configuring the `artemis-odb-maven-plugin`.
  - New method `Entity#createComponent(Class<Component>)`.
  - Annotation processor validates pooled and packed component types.
+ - Managers support `@Mapper` annotation.
  - Fix: DelayedEntityProcessingSystem prematurely expired entities.
  - Fix: Recycled entities would sometimes have their components cleared when
    recycled during the same round as the original entity was deleted.